Move common components to components dir=
